Added interface for calculator

+++ Calculator/src/ChildCalculator.java
@@ -7,25 +7,25 @@ public class ChildCalculator extends Calculator {
     }
 
     @Override
-    public int add(int x, int y) {
+    public double add(double x, double y) {
         System.out.println("Overriding Add");
         return super.add(x, y);
     }
 
     @Override
-    public int subtract(int x, int y) {
+    public double subtract(double x, double y) {
         System.out.println("Overriding Subtract");
         return super.subtract(x, y);
     }
 
     @Override
-    public int divide(int x, int y) {
+    public double divide(double x, double y) {
         System.out.println("Overriding Divide");
         return super.divide(x, y);
     }
 
     @Override
-    public int multiply(int x, int y) {
+    public double multiply(double x, double y) {
         System.out.println("Overriding Multiply");
         return super.multiply(x, y);
     }
